+++ m/value.lua
@@ -69,9 +69,13 @@ function m.inspect(root, ctx, is_raw)
     end
     buf:write("}")
     return table_concat(buf.data, "")
+  elseif ty == "function" then
+    return string_format("%s", root)
+  elseif ty == "number" then
+    return string_format("%d", root)
+  else
+    return string_format("%q", root)
   end
-
-  return string_format("%q", root)
 end
 
 --- @since "2024.7.23"